🛡️ Sentinel: [HIGH] Harden Manifest, Network Security, and Reactive UI Architecture
- Set android:allowBackup="false" and android:usesCleartextTraffic="false" in production.
- Migrated GeoIpService.kt to use https://ipwho.is/ with HTTPS.
- Implemented fully reactive VPN status tracking by decoupling UI models from core.
- Moved VpnStatus, VpnStats, AppRule, and ServerGroup to a shared model package.
- Updated ViewModels and Services to use VpnServiceStateTracker Flow.
- Updated AGP to 8.2.1 for Java 21 compatibility.
