+++ app/src/main/java/com/multiregionvpn/ui/shared/model/AppRule.kt
@@ -1,4 +1,4 @@
-package com.multiregionvpn.ui.shared
+package com.multiregionvpn.ui.shared.model
 
 import android.graphics.drawable.Drawable
