+++ app/src/main/java/com/multiregionvpn/ui/shared/model/VpnStats.kt
@@ -1,0 +1,11 @@
+package com.multiregionvpn.ui.shared.model
+
+/**
+ * Shared VPN statistics model
+ */
+data class VpnStats(
+    val bytesSent: Long = 0L,
+    val bytesReceived: Long = 0L,
+    val connectionTimeSeconds: Long = 0L,
+    val activeConnections: Int = 0
+)
