+++ app/src/main/java/com/multiregionvpn/ui/shared/model/VpnStatus.kt
@@ -1,4 +1,4 @@
-package com.multiregionvpn.ui.shared
+package com.multiregionvpn.ui.shared.model
 
 /**
  * Shared VPN status enum used by both Mobile and TV UIs
@@ -9,4 +9,3 @@ enum class VpnStatus(val displayText: String) {
     DISCONNECTED("Disconnected"),
     ERROR("Error")
 }
-
